[repository #190] prepare reference iteration
