+++ git-repository/src/easy/ext/reference.rs
@@ -144,6 +144,18 @@ pub trait ReferenceAccessExt: easy::Access + Sized {
             .ok_or(reference::find::existing::Error::NotFound)
     }
 
+    fn iter_references(&self) -> Result<easy::iter::references::State<'_, Self>, easy::iter::references::Error> {
+        let _state = self.state();
+        todo!("")
+        // Ok(easy::iter::References {
+        //     inner: self
+        //         .repo()?
+        //         .refs
+        //         .iter(state.assure_packed_refs_uptodate(&repo.refs)?.as_ref())?,
+        //     access: self,
+        // })
+    }
+
     fn try_find_reference<'a, Name, E>(&self, name: Name) -> Result<Option<Reference<'_, Self>>, reference::find::Error>
     where
         Name: TryInto<PartialNameRef<'a>, Error = E>,
