+++ git-repository/src/easy/mod.rs
@@ -32,6 +32,7 @@ pub(crate) mod ext;
 pub mod borrow;
 pub mod commit;
 pub mod head;
+pub mod iter;
 pub mod object;
 mod oid;
 pub mod reference;
