+++ git-repository/tests/easy/ext/mod.rs
@@ -1,2 +1,2 @@
-mod head;
 mod object;
+mod reference;
